requirements.txt
captum @ file:///home/conda/feedstock_root/build_artifacts/captum_1676431791267/work
causal-conv1d==1.0.0
certifi==2024.2.2
cffi @ file:///croot/cffi_1714483155441/work
chardet==5.2.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.6
contourpy==1.1.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
einops==0.7.0
exceptiongroup==1.2.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
joblib==1.4.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader==0.4
lightning-utilities==0.11.2
lit==18.1.4
llvmlite==0.41.1
mamba-ssm @ file:///ML/mamba_ssm-1.0.1%2Bcu118torch1.13cxx11abiFALSE-cp38-cp38-linux_x86_64.whl#sha256=31f374402f85b8b8ea5d62771aa9fd293fd0447faf10c124090620960df2611c
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1610127565888/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091694890/work
MedPy==0.5.1
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1667391556261/work
ninja==1.11.1.1
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
packaging==24.0
pandas==2.0.3
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pretty-errors==1.2.25
protobuf==5.26.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1709721012883/work
pytest==8.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytoolconfig==1.3.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rope==1.13.0
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shap==0.44.1
SimpleITK==2.3.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicer==0.0.7
submitit==1.5.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchinfo==1.8.0
torchmetrics==1.4.0
torchvision==0.14.0+cu117
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.2
transformers==4.40.1
triton==2.0.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
urllib3==2.2.1
yacs==0.1.8
zipp==3.18.1
